// src/types/apis/userProfile/get_fetchWidgetTheme.ts
export interface get_fetchWidgetTheme {
  brand_primary_color: string;
  brand_secondary_color: null;
  text_primary_color: null;
  text_secondary_color: null;
  button_primary_text_color: null;
  layout_container_background: null;
  layout_card_background: null;
  layout_element_border: null;
  layout_container_border: null;
  advanced_font_family: null;
  advanced_border_radius: null;
  advanced_shadow_style: null;
}